move int to size_t for primes

+++ Ass1/utils.hpp
@@ -3,7 +3,7 @@
 #include <math.h>
 
 // Calculation
-int gcd(int a, int b, int& flop) {
+size_t gcd(size_t a, size_t b, size_t& flop) {
    if (a == 0 || b == 0) {
         flop += 2;
         return 0;
@@ -23,13 +23,13 @@ int gcd(int a, int b, int& flop) {
 }
 
 // calculate the number based on its index in a cyclically distributed boolean array
-int index_to_number(int i, int pid, int p, int &flop) {
+size_t index_to_number(size_t i, size_t pid, size_t p, size_t &flop) {
     flop += 3;
     return pid + i * p + 1;
 }
 
 // calculate the index of the number in a cyclically distirbuted boolean array
-int number_to_index(int number, int p, int pid, int &flop) {
+size_t number_to_index(size_t number, size_t p, size_t pid, size_t &flop) {
     // return static_cast<int>(round(number / (p * 1.0)));
     return ceil((number - 1 - pid) / (p * 1.0));
 }
@@ -37,7 +37,7 @@ int number_to_index(int number, int p, int pid, int &flop) {
 
 // Communication
 // function that sends a number to all the other processors
-void put_numbers_to_all(int number, int pid, int p, bulk::queue<int> &q) {
+void put_numbers_to_all(size_t number, int pid, int p, bulk::queue<size_t> &q) {
     for (int i = 0; i < p; i ++) {
         if (i != pid) { 
             q(i).send(number);
@@ -46,9 +46,9 @@ void put_numbers_to_all(int number, int pid, int p, bulk::queue<int> &q) {
 }
 
 // function that sends the locally found primes to only the processors that might have multiples of it
-void put_prime_multiples(int prime, int n, int p, bulk::queue<int> &q, int& flop, bulk::world& world) {
+void put_prime_multiples(size_t prime, size_t n, int p, bulk::queue<size_t> &q, size_t& flop, bulk::world& world) {
     flop ++;
-    for (int i = prime * 2; i <= n; i += prime) {
+    for (size_t i = prime * 2; i <= n; i += prime) {
         // world.log("send prime %d to processor %d", prime, (i - 1) % p);
         q((i - 1) % p).send(i);
         flop += 3;
